Refactors, move name patterns
