+++ deduce/annotate.py
@@ -1,22 +1,20 @@
 """ The annotate module contains the code for annotating text"""
 import re
 from dataclasses import dataclass
-from typing import Union, Optional, Any
-
-from abc import ABC, abstractmethod
-
-import itertools
+from typing import Union
 
 import docdeid
-from docdeid.annotation.annotation_processor import OverlapResolver
-from docdeid.annotation.annotator import RegexpAnnotator, MultiTokenLookupAnnotator
-from docdeid.datastructures.lookup import LookupList
+from deduce.name_patterns import PrefixWithNamePattern, InterfixWithNamePattern, InitialWithCapitalPattern, \
+    InitiaalInterfixCapitalPattern, FirstNameLookupPattern, SurnameLookupPattern, PersonFirstNamePattern, \
+    PersonInitialFromNamePattern, PersonSurnamePattern, PersonInitialsPattern, PersonGivenNamePattern
+from docdeid.annotate.annotation_processor import OverlapResolver
+from docdeid.annotate.annotator import RegexpAnnotator, MultiTokenLookupAnnotator
+from docdeid.ds.lookup import LookupList
 from docdeid.str.processor import LowercaseString
-from nltk.metrics import edit_distance
 
 from deduce import utility
 from deduce.lookup.lookup_lists import get_lookup_lists
-from deduce.tokenizer import Tokenizer
+from deduce.tokenizer import Tokenizer, TokenContext
 
 
 def _initialize():
@@ -44,424 +42,22 @@ class Person:
     given_name: str = None
 
 
-class TokenContext:
-
-    def __init__(self, position: int, tokens: list[docdeid.Token]):
-
-        self._tokens = tokens
-        self._position = position
-
-    @property
-    def token(self):
-        return self._tokens[self._position]
-
-    def next(self, num: int = 1):
-
-        current_token = self._tokens[self._position]
-
-        for _ in range(num):
-            current_token = self._get_next_token(current_token.index)
-
-            if current_token is None:
-                return None
-
-        return current_token
-
-    def previous(self, num: int = 1):
-
-        current_token = self._tokens[self._position]
-
-        for _ in range(num):
-            current_token = self.get_previous_token(current_token.index)
-
-            if current_token is None:
-                return None
-
-        return current_token
-
-    def num_tokens_from_position(self):
-
-        return len(self._tokens) - self._position
-
-    def get_token(self, pos: int):
-        return self._tokens[pos]
-
-    def get_token_at_num_from_position(self, num=0):
-        return self._tokens[self._position + num]
-
-    def _get_next_token(self, i: int) -> Union[docdeid.Token, None]:
-
-        if i == len(self._tokens):
-            return None
-
-        for token in self._tokens[i + 1:]:
-
-            if (
-                    token.text[0] == ")"
-                    or token.text[0] == ">"
-                    or utility.any_in_text(["\n", "\r", "\t"], token.text)
-            ):
-                return None
-
-            if token.text[0].isalpha():
-                return token
-
-        return None
-
-    def get_previous_token(self, i: int) -> Union[docdeid.Token, None]:
-
-        if i == 0:
-            return None
-
-        for token in self._tokens[i - 1:: -1]:
-
-            if (
-                    token.text[0] == "("
-                    or token.text[0] == "<"
-                    or utility.any_in_text(["\n", "\r", "\t"], token.text)
-            ):
-                return None
-
-            if token.text[0].isalpha():
-                return token
-
-        return None
-
-    @property
-    def next_token(self):
-        return self.next(1)
-
-    @property
-    def previous_token(self):
-        return self.previous(1)
-
-
-class TokenContextPattern(ABC):
-
-    def __init__(self, tag: str):
-        self._tag = tag
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-        return True
-
-    @abstractmethod
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> Union[bool, tuple[bool, Any]]:
-        pass
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-        return token_context.token, token_context.token, self._tag
-
-    def apply(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> Union[None, tuple]:
-
-        if not self.precondition(token_context, meta_data):
-            return None
-
-        match = self.match(token_context, meta_data)
-        match_info = None
-
-        if isinstance(match, tuple):
-            match, match_info = match  # unpack
-
-        if match:
-            return self.annotate(token_context, match_info)
-
-
-class PrefixWithNamePattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return token_context.next_token is not None
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                token_context.token.text.lower() in _lookup_lists["prefixes"],
-                token_context.next_token.text[0].isupper(),
-                token_context.next_token.text.lower() not in _lookup_lists["whitelist"],
-            ]
-        )
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return token_context.token, token_context.next_token, self._tag
-
-
-class InterfixWithNamePattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-        return token_context.next_token is not None
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                token_context.token.text.lower() in _lookup_lists["interfixes"],
-                token_context.next_token.text in _lookup_lists["interfix_surnames"],
-                token_context.next_token.text.lower() not in _lookup_lists["whitelist"],
-            ]
-        )
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return token_context.token, token_context.next_token, self._tag
-
-
-class InitialWithCapitalPattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return token_context.next_token is not None
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                token_context.token.text[0].isupper(),
-                len(token_context.token.text) == 1,
-
-                len(token_context.next_token.text) > 3,
-                token_context.next_token.text[0].isupper(),
-                token_context.next_token.text.lower() not in _lookup_lists["whitelist"],
-            ]
-        )
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return token_context.token, token_context.next_token, self._tag
-
-
-class InitiaalInterfixCapitalPattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                token_context.previous_token is not None,
-                token_context.next_token is not None,
-            ]
-        )
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                token_context.previous_token.text[0].isupper(),
-                len(token_context.previous_token.text) == 1,
-
-                token_context.token.text in _lookup_lists["interfixes"],
-
-                token_context.next_token.text[0].isupper(),
-            ]
-        )
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return token_context.previous_token, token_context.next_token, self._tag
-
-
-class FirstNameLookupPattern(TokenContextPattern):
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                token_context.token.text in _lookup_lists["first_names"],
-                token_context.token.text.lower() not in _lookup_lists["whitelist"],
-            ]
-        )
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return token_context.token, token_context.token, self._tag
-
-
-class SurnameLookupPattern(TokenContextPattern):
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-        return all(
-            [
-                token_context.token.text in _lookup_lists["surnames"],
-                token_context.token.text.lower() not in _lookup_lists["whitelist"],
-            ]
-        )
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-        return token_context.token, token_context.token, self._tag
-
-
-class PersonFirstNamePattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                meta_data is not None,
-                'person' in meta_data,
-                meta_data['person'].first_names is not None
-            ]
-        )
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> Union[bool, tuple[bool, Any]]:
-
-        for i, first_name in enumerate(meta_data['person'].first_names):
-
-            condition = token_context.token.text == first_name or \
-                all(
-                    [
-                        len(token_context.token.text) > 3,
-                        edit_distance(token_context.token.text, first_name, transpositions=True) <= 1
-                    ]
-                )
-
-            if condition:
-                return True, token_context.token.index
-
-        return False
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return (
-            token_context.get_token(pos=match_info),
-            token_context.get_token(pos=match_info),
-            self._tag
-        )
-
-
-class PersonInitialFromNamePattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                meta_data is not None,
-                'person' in meta_data,
-                meta_data['person'].first_names is not None
-            ]
-        )
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> Union[bool, tuple[bool, Any]]:
-
-        for i, first_name in enumerate(meta_data['person'].first_names):
-
-            if token_context.token.text == first_name[0]:
-
-                next_token = token_context.next(1)
-                if (next_token is not None) and next_token == ".":
-                    return True, (token_context.token.index, next_token.index)
-
-                return True, (token_context.token.index, token_context.token.index)
-
-        return False
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return (
-            token_context.get_token(pos=match_info[0]),
-            token_context.get_token(pos=match_info[1]),
-            self._tag
-        )
-
-
-class PersonSurnamePattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                meta_data is not None,
-                'person' in meta_data,
-                meta_data['person'].surname is not None
-            ]
-        )
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> Union[bool, tuple[bool, Any]]:
-
-        surname_pattern = [token.text for token in tokenizer.tokenize(meta_data['patient_surname'])]
-
-        if len(surname_pattern) > token_context.num_tokens_from_position():
-            return False
-
-        return all(
-                [
-                    edit_distance(surname_token, token_context.get_token_at_num_from_position(i).text, transpositions=True) <= 1
-                    for surname_token, i in zip(surname_pattern, itertools.count())
-                ]
-            ), len(surname_pattern)-1
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return (
-            token_context.get_token_at_num_from_position(0),
-            token_context.get_token_at_num_from_position(match_info),
-            self._tag,
-        )
-
-
-class PersonInitialsPattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                meta_data is not None,
-                'person' in meta_data,
-                meta_data['person'].initials is not None
-            ]
-        )
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return token_context.token.text == meta_data['person'].initials
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return token_context.token, token_context.token, self._tag
-
-
-class PersonGivenNamePattern(TokenContextPattern):
-
-    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return all(
-            [
-                meta_data is not None,
-                'person' in meta_data,
-                meta_data['person'].given_name is not None
-            ]
-        )
-
-    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
-
-        return (token_context.token.text == meta_data['person'].given_name) or \
-               all(
-                   [
-                       len(token_context.token.text) > 3,
-                       edit_distance(token_context.token.text, meta_data['person'].given_name, transpositions=True) <= 1,
-                   ]
-               )
-
-    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
-
-        return token_context.token, token_context.token, self._tag
-
-
 class NamesAnnotator(docdeid.BaseAnnotator):
 
     def __init__(self):
 
         self._patterns = [
-            PrefixWithNamePattern(tag="prefix+naam"),
-            InterfixWithNamePattern(tag="interfix+naam"),
-            InitialWithCapitalPattern(tag="initiaal+naam"),
-            InitiaalInterfixCapitalPattern(tag="initiaal+interfix+naam"),
-            FirstNameLookupPattern(tag="voornaam_onbekend"),
-            SurnameLookupPattern(tag="achternaam_onbekend"),
+            PrefixWithNamePattern(tag="prefix+naam", lookup_lists=_lookup_lists),
+            InterfixWithNamePattern(tag="interfix+naam", lookup_lists=_lookup_lists),
+            InitialWithCapitalPattern(tag="initiaal+naam", lookup_lists=_lookup_lists),
+            InitiaalInterfixCapitalPattern(tag="initiaal+interfix+naam", lookup_lists=_lookup_lists),
+            FirstNameLookupPattern(tag="voornaam_onbekend", lookup_lists=_lookup_lists),
+            SurnameLookupPattern(tag="achternaam_onbekend", lookup_lists=_lookup_lists),
             PersonFirstNamePattern(tag="voornaam_patient"),
             PersonInitialFromNamePattern(tag="initiaal_patient"),
             PersonInitialsPattern(tag="initialen_patient"),
             PersonGivenNamePattern(tag="roepnaam_patient"),
-            PersonSurnamePattern(tag="achternaam_patient")
+            PersonSurnamePattern(tag="achternaam_patient", tokenizer=tokenizer)
         ]
 
     @staticmethod
