+++ deduce/str/processor.py
@@ -1,7 +1,7 @@
 import re
 from typing import Optional
 
-from docdeid.datastructures import LookupList
+from docdeid.ds import LookupList
 from docdeid.str.processor import BaseStringProcessor, LowercaseString
