+++ deduce/tokenizer.py
@@ -1,11 +1,13 @@
 """ This module contains all tokenizing functionality """
 import itertools
+from abc import ABC, abstractmethod
 from enum import Enum, auto
-from typing import Iterable
+from typing import Iterable, Union, Optional, Any
 
 import docdeid
-import docdeid.tokenizer.tokenizer
-from docdeid.datastructures.lookup import LookupTrie
+import docdeid.tokenize.tokenizer
+from deduce import utility
+from docdeid.ds.lookup import LookupTrie
 
 
 class _CharType(Enum):
@@ -120,3 +122,125 @@ class Tokenizer(docdeid.BaseTokenizer):
             )
             for token, i in zip(tokens, itertools.count())
         ]
+
+
+class TokenContext:
+
+    def __init__(self, position: int, tokens: list[docdeid.Token]):
+
+        self._tokens = tokens
+        self._position = position
+
+    @property
+    def token(self):
+        return self._tokens[self._position]
+
+    def next(self, num: int = 1):
+
+        current_token = self._tokens[self._position]
+
+        for _ in range(num):
+            current_token = self._get_next_token(current_token.index)
+
+            if current_token is None:
+                return None
+
+        return current_token
+
+    def previous(self, num: int = 1):
+
+        current_token = self._tokens[self._position]
+
+        for _ in range(num):
+            current_token = self.get_previous_token(current_token.index)
+
+            if current_token is None:
+                return None
+
+        return current_token
+
+    def num_tokens_from_position(self):
+
+        return len(self._tokens) - self._position
+
+    def get_token(self, pos: int):
+        return self._tokens[pos]
+
+    def get_token_at_num_from_position(self, num=0):
+        return self._tokens[self._position + num]
+
+    def _get_next_token(self, i: int) -> Union[docdeid.Token, None]:
+
+        if i == len(self._tokens):
+            return None
+
+        for token in self._tokens[i + 1:]:
+
+            if (
+                    token.text[0] == ")"
+                    or token.text[0] == ">"
+                    or utility.any_in_text(["\n", "\r", "\t"], token.text)
+            ):
+                return None
+
+            if token.text[0].isalpha():
+                return token
+
+        return None
+
+    def get_previous_token(self, i: int) -> Union[docdeid.Token, None]:
+
+        if i == 0:
+            return None
+
+        for token in self._tokens[i - 1:: -1]:
+
+            if (
+                    token.text[0] == "("
+                    or token.text[0] == "<"
+                    or utility.any_in_text(["\n", "\r", "\t"], token.text)
+            ):
+                return None
+
+            if token.text[0].isalpha():
+                return token
+
+        return None
+
+    @property
+    def next_token(self):
+        return self.next(1)
+
+    @property
+    def previous_token(self):
+        return self.previous(1)
+
+
+class TokenContextPattern(ABC):
+
+    def __init__(self, tag: str):
+        self._tag = tag
+
+    def precondition(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> bool:
+        return True
+
+    @abstractmethod
+    def match(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> Union[bool, tuple[bool, Any]]:
+        pass
+
+    def annotate(self, token_context: TokenContext, match_info=None) -> tuple:
+        return token_context.token, token_context.token, self._tag
+
+    def apply(self, token_context: TokenContext, meta_data: Optional[dict] = None) -> Union[None, tuple]:
+
+        if not self.precondition(token_context, meta_data):
+            return None
+
+        match = self.match(token_context, meta_data)
+        match_info = None
+
+        if isinstance(match, tuple):
+            match, match_info = match  # unpack
+
+        if match:
+            return self.annotate(token_context, match_info)
